requirements.txt
google-genai
opencv-contrib-python
numpy
Pillow
pvporcupine
vosk
sounddevice
pygame
pyyaml
python-dotenv
rapidfuzz
Sherpa-ONNX